Adjust language icon UI to the new icon system
