+++ client/common/src/index.ts
@@ -1,3 +1,4 @@
 export * from './types'
 export * from './errors'
 export * from './util'
+export * from './languages'
