+++ client/shared/src/backend/repo.ts
@@ -75,6 +75,7 @@ export const fetchTreeEntries = memoizeObservable(
                     isRoot
                     url
                     entries(first: $first) {
+                        __typename
                         ...TreeEntryFields
                         ... on GitBlob {
                             languages
