+++ client/web-sveltekit/src/lib/wildcard/index.ts
@@ -10,3 +10,4 @@ export { default as MenuLink } from './menu/MenuLink.svelte'
 export { default as MenuSeparator } from './menu/MenuSeparator.svelte'
 export { default as Submenu } from './menu/Submenu.svelte'
 export { default as MenuRadioGroup } from './menu/MenuRadioGroup.svelte'
+export { getFileIconInfo } from '@sourcegraph/wildcard/src/components/Icon'
