+++ client/web/src/search/suggestion/DidYouMean.tsx
@@ -3,9 +3,9 @@ import React, { useEffect, useMemo } from 'react'
 import { mdiArrowRight } from '@mdi/js'
 
 import { SyntaxHighlightedSearchQuery } from '@sourcegraph/branded'
+import { ALL_LANGUAGES } from '@sourcegraph/common'
 import type { SearchPatternType } from '@sourcegraph/shared/src/graphql-operations'
 import type { CaseSensitivityProps, SearchPatternTypeProps, SearchContextProps } from '@sourcegraph/shared/src/search'
-import { ALL_LANGUAGES } from '@sourcegraph/shared/src/search/query/languageFilter'
 import { stringHuman } from '@sourcegraph/shared/src/search/query/printer'
 import { scanSearchQuery } from '@sourcegraph/shared/src/search/query/scanner'
 import { createLiteral, type Pattern, type Token } from '@sourcegraph/shared/src/search/query/token'
