+++ client/wildcard/src/components/Icon/index.ts
@@ -1,2 +1,3 @@
 export * from './Icon'
 export * from './language-icon/LanguageIcon'
+export { getFileIconInfo } from './language-icon/language-icons'
