+++ client/wildcard/src/components/index.ts
@@ -55,7 +55,7 @@ export { Panel } from './Panel'
 export { Tooltip, TooltipOpenChangeReason } from './Tooltip'
 export { Tree, flattenTree } from './Tree'
 export { Card, CardBody, CardHeader, CardList, CardSubtitle, CardText, CardTitle, CardFooter } from './Card'
-export { Icon, LanguageIcon, FILE_ICONS, FileExtension } from './Icon'
+export { Icon, LanguageIcon, getFileIconInfo } from './Icon'
 export { ButtonLink } from './ButtonLink'
 export { Menu, MenuButton, MenuDivider, MenuHeader, MenuItem, MenuLink, MenuList, MenuText } from './Menu'
 export { NavMenu } from './NavMenu'
